Chore/vrf 526 add ctf test vrfv2plus (#10262)
* fix VRF V3 setup script

* VRF-526: add VRF V2 Plus CTF smoke test

* VRF-526: removing comments

* VRF-526: fixing solidity lint issue

* Apply suggestions from code review


* VRF-526: PR review changes

* VRF-526: minor fix

---------
