+++ .github/workflows/integration-tests.yml
@@ -274,6 +274,10 @@ jobs:
             nodes: 1
             os: ubuntu20.04-8cores-32GB
             pyroscope_env: ci-smoke-vrf2-evm-simulated
+          - name: vrfv2plus
+            nodes: 1
+            os: ubuntu-latest
+            pyroscope_env: ci-smoke-vrf2plus-evm-simulated
           - name: forwarder_ocr
             nodes: 1
             os: ubuntu20.04-8cores-32GB
@@ -464,7 +468,7 @@ jobs:
           hostname: ${{ secrets.GRAFANA_CLOUD_HOST }}
           this-job-name: ETH Smoke Tests
         continue-on-error: true
-  
+
   # Run the setup if the matrix finishes but this time save the cache if we have a cache hit miss
   # this will also only run if both of the matrix jobs pass
   eth-smoke-go-mod-cache:
